Avoid removing $HOME by clean.sh
If LOGDIR is not set, the default value is LOGDIR="${LOGFILE%/*}". When
LOGFILE=$HOME/log.file LOGDIR become equal to $HOME directory.
When clean.sh is called $LOGDIR is removed automatically.
This patch ignores removing $LOGDIR if it is equal to $HOME.

Closes-Bug: #1526433

+++ clean.sh
@@ -124,10 +124,10 @@ sudo rm -rf $DATA_DIR $DEST/status
 if [[ -n "$LOGFILE" ]] && [[ -f "$LOGFILE" ]]; then
     sudo rm -f $LOGFILE
 fi
-if [[ -n "$LOGDIR" ]] && [[ -d "$LOGDIR" ]]; then
+if [[ -n "$LOGDIR" ]] && [[ -d "$LOGDIR" ]] && [[ "$LOGDIR" != "$HOME" ]]; then
     sudo rm -rf $LOGDIR
 fi
-if [[ -n "$SCREEN_LOGDIR" ]] && [[ -d "$SCREEN_LOGDIR" ]]; then
+if [[ -n "$SCREEN_LOGDIR" ]] && [[ -d "$SCREEN_LOGDIR" ]] && [[ "$SCREEN_LOGDIR" != "$HOME" ]]; then
     sudo rm -rf $SCREEN_LOGDIR
 fi
